be/pyproject.toml
[tool.poetry]
name = "be"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12.0,<3.13.0"
django = "4.2.20"
djangorestframework = "^3.16.0"

[tool.poetry.requires-plugins]
poetry-plugin-shell = ">=1.0.1"
poetry-plugin-export = ">=1.8"